added checkbox to sitl_gui

+++ sitl_gui.py
@@ -1,3 +1,5 @@
+#!/home/r32401vc/anaconda3/bin/python3
+
 import sys
 import subprocess
 import mavsdk
@@ -9,6 +11,9 @@ from tkinter import filedialog
 from tkinter import messagebox
 from tkinter import *
 import paho.mqtt.client as mqtt
+
+# Global Variables
+gazebo = None
 
 # this is the function called when the button is clicked
 def StartClickFunction():
@@ -37,15 +42,18 @@ def LaunchClickFunction():
     global gazebo
     gazebo = subprocess.Popen(
         ["bash", "gazebo_sitl_multiple_run.sh", "-n", str(no_drones)],
-        # capture_output=True,
-        # text=True,
         cwd=path_label.cget("text") + "/Tools/",
         stdin=None,
         stdout=None,
         stderr=None,
-    )
+        preexec_fn=os.setsid,
+    )  # last argument important to allow process to be killed
+
     global mavsdk_server
     mavsdk_server = [None] * no_drones
+    global script
+    script = [None] * no_drones
+
     for i in range(0, no_drones):
         mavsdk_server[i] = subprocess.Popen(
             [
@@ -55,17 +63,12 @@ def LaunchClickFunction():
                 str(50041 + i),
                 "udp://:" + str(14540 + i),
             ],
-            # capture_output=True,
-            # text=True,
             cwd=os.path.split(mavsdk.__file__)[0] + "/bin",
             stdin=None,
             stdout=None,
             stderr=None,
         )
-    global script
-    script = [None] * no_drones
-    for i in range(0, no_drones):
-        print(i)
+
         script[i] = subprocess.Popen(
             [
                 sys.executable,
@@ -74,12 +77,7 @@ def LaunchClickFunction():
                 str(no_drones),
                 str(50041 + i),
             ],
-            # capture_output=True,
-            # shell=True,
             cwd=os.getcwd(),
-            # stdin=None,
-            # stdout=None,
-            # stderr=None,
         )
 
 
@@ -89,6 +87,22 @@ def SelectClickFunction():
     path_label.config(text=firmware_path)
     with open("config.txt", "w") as f:
         f.write(firmware_path)
+
+
+def SITLCheckFunction():
+    print("function activated")
+    print(check_var.get())
+    if check_var.get() == 0:
+        launch_btn.grid_remove()
+        select_firmware_btn.grid_remove()
+        path_label.grid_remove()
+        no_drones_label.grid_remove()
+    else:
+        print("returning items")
+        launch_btn.grid()
+        select_firmware_btn.grid()
+        path_label.grid()
+        no_drones_label.grid()
 
 
 # sends mqtt commands to broker
@@ -111,36 +125,38 @@ def validate_int(input):
 
 
 def on_closing():
-    sig = signal.SIGKILL
-    os.kill(gazebo.pid, sig)
-    for i in range(0, len(script)):
-        os.kill(mavsdk_server[i].pid, sig)
-        os.kill(script[i].pid, sig)
+    if gazebo != None:
+        sig = signal.SIGTERM
+        os.killpg(os.getpgid(gazebo.pid), sig)
+        for i in range(0, len(script)):
+            os.kill(mavsdk_server[i].pid, sig)
+            os.kill(script[i].pid, sig)
     root.destroy()
 
 
 root = Tk()
-
+check_var = IntVar(value=1)
 # This is the section of code which creates the main window
 root.geometry("636x525")
 root.configure(background="#F0F8FF")
-root.title("Swarm Experiment")
 
-Button(
+launch_btn = Button(
     root,
     text="Launch SITL Simulation",
     bg="#008B45",
     font=("arial", 12, "normal"),
     command=LaunchClickFunction,
-).grid(row=0, column=0)
+)
+launch_btn.grid(row=0, column=0)
 
-Button(
+select_firmware_btn = Button(
     root,
     text="Select PX4 Firmware",
     # bg="#008B45",
     font=("arial", 12, "normal"),
     command=SelectClickFunction,
-).grid(row=1, column=0)
+)
+select_firmware_btn.grid(row=1, column=0)
 
 Button(
     root,
@@ -174,7 +190,18 @@ Button(
     command=LandClickFunction,
 ).grid(row=2, column=3)
 
-Label(root, text="Number of Drones:").grid(row=0, column=1)
+sitl_check = Checkbutton(
+    root,
+    text="SITL",
+    variable=check_var,
+    onvalue=1,
+    offvalue=0,
+    command=SITLCheckFunction,
+)
+sitl_check.grid(row=3, column=0)
+
+no_drones_label = Label(root, text="Number of Drones:")
+no_drones_label.grid(row=0, column=1)
 
 path_label = Label(root)
 path_label.grid(row=1, column=1)
@@ -186,5 +213,7 @@ drone_no_entry = Entry(root)
 drone_no_entry.grid(row=0, column=2)
 drone_no_entry.insert(END, "3")
 
+root.title("Cascade Demo")
+root.iconphoto(True, tk.PhotoImage(file="img/cascade-logo.png"))
 root.protocol("WM_DELETE_WINDOW", on_closing)
 root.mainloop()
